refactor: API 경로 및 쿼리 관련 파일 구조 정리
