+++ apps/client/src/shared/apis/prompt/query-key.ts
