+++ apps/client/src/shared/apis/prompt/type.ts
