+++ apps/client/src/shared/apis/user/type.ts
